[Enterprise Search] Refactor shared SchemaAddFieldModal component (#99096)
* Move SchemaAddFieldModal to its own folder

+ misc test cleanup - remove unnecessary jest.spyOn in individual tests, it's already in the beforeEach

* i18n fixes

- Move constants to subfolder
- Fix various i18n IDs/var names - modal titles is not a fieldNote title
- Fix an i18n string that should be a FormattedMessage (could have grammar issues otherwise)
- Add missing i18n strings - labels (shared, AS will also use these as column headers) & placeholder
- Import order

* Move formatFieldName util to its own file

- simplify leading/trailing trimming regex
- add unit tests - primarily for documenting regexes & providing examples

* Refactor modal form submission

- See https://elastic.github.io/eui/#/layout/modal#forms-in-a-modal for documentation - form should be submitted via ID & EuiModalFooter form={id}

* Misc props cleanup

- Move optional props to bottom of type list
- Remove unnecessary props (default EUI behavior)
- Data test subj cleanup
- Add missing `disabled` passed prop, add unit test for disableForm prop
