+++ x-pack/plugins/enterprise_search/public/applications/shared/schema/constants.ts
@@ -7,40 +7,13 @@
 
 import { i18n } from '@kbn/i18n';
 
-export const FIELD_NAME_CORRECT_NOTE = i18n.translate(
-  'xpack.enterpriseSearch.schema.addFieldModal.fieldNameNote.correct',
-  {
-    defaultMessage: 'Field names can only contain lowercase letters, numbers, and underscores',
-  }
-);
+export const FIELD_NAME = i18n.translate('xpack.enterpriseSearch.schema.fieldNameLabel', {
+  defaultMessage: 'Field name',
+});
 
-export const FIELD_NAME_CORRECTED_PREFIX = i18n.translate(
-  'xpack.enterpriseSearch.schema.addFieldModal.fieldNameNote.corrected',
-  {
-    defaultMessage: 'The field will be named',
-  }
-);
-
-export const FIELD_NAME_MODAL_TITLE = i18n.translate(
-  'xpack.enterpriseSearch.schema.addFieldModal.fieldNameNote.title',
-  {
-    defaultMessage: 'Add a New Field',
-  }
-);
-
-export const FIELD_NAME_MODAL_DESCRIPTION = i18n.translate(
-  'xpack.enterpriseSearch.schema.addFieldModal.fieldNameNote.description',
-  {
-    defaultMessage: 'Once added, a field cannot be removed from your schema.',
-  }
-);
-
-export const FIELD_NAME_MODAL_ADD_FIELD = i18n.translate(
-  'xpack.enterpriseSearch.schema.addFieldModal.fieldNameNote.addField',
-  {
-    defaultMessage: 'Add field',
-  }
-);
+export const FIELD_TYPE = i18n.translate('xpack.enterpriseSearch.schema.fieldTypeLabel', {
+  defaultMessage: 'Field type',
+});
 
 export const ERROR_TABLE_ID_HEADER = i18n.translate(
   'xpack.enterpriseSearch.schema.errorsTable.heading.id',
