+++ x-pack/plugins/translations/translations/ja-JP.json
@@ -7607,11 +7607,11 @@
     "xpack.enterpriseSearch.overview.subheading": "開始する製品を選択します。",
     "xpack.enterpriseSearch.productName": "エンタープライズサーチ",
     "xpack.enterpriseSearch.readOnlyMode.warning": "エンタープライズ サーチは読み取り専用モードです。作成、編集、削除などの変更を実行できません。",
-    "xpack.enterpriseSearch.schema.addFieldModal.fieldNameNote.addField": "フィールドの追加",
+    "xpack.enterpriseSearch.schema.addFieldModal.addFieldButtonLabel": "フィールドの追加",
+    "xpack.enterpriseSearch.schema.addFieldModal.description": "追加すると、フィールドはスキーマから削除されます。",
+    "xpack.enterpriseSearch.schema.addFieldModal.title": "新しいフィールドを追加",
     "xpack.enterpriseSearch.schema.addFieldModal.fieldNameNote.correct": "フィールド名には、小文字、数字、アンダースコアのみを使用できます。",
-    "xpack.enterpriseSearch.schema.addFieldModal.fieldNameNote.corrected": "フィールド名が変更されます",
-    "xpack.enterpriseSearch.schema.addFieldModal.fieldNameNote.description": "追加すると、フィールドはスキーマから削除されます。",
-    "xpack.enterpriseSearch.schema.addFieldModal.fieldNameNote.title": "新しいフィールドを追加",
+    "xpack.enterpriseSearch.schema.addFieldModal.fieldNameNote.corrected": "フィールド名が変更されます {correctedName}",
     "xpack.enterpriseSearch.schema.errorsCallout.buttonLabel": "エラーを表示",
     "xpack.enterpriseSearch.schema.errorsTable.control.review": "見直し",
     "xpack.enterpriseSearch.schema.errorsTable.heading.error": "エラー",
