+++ x-pack/plugins/translations/translations/zh-CN.json
@@ -7676,11 +7676,11 @@
     "xpack.enterpriseSearch.overview.subheading": "选择产品开始使用。",
     "xpack.enterpriseSearch.productName": "企业搜索",
     "xpack.enterpriseSearch.readOnlyMode.warning": "企业搜索处于只读模式。您将无法执行更改，例如创建、编辑或删除。",
-    "xpack.enterpriseSearch.schema.addFieldModal.fieldNameNote.addField": "添加字段",
+    "xpack.enterpriseSearch.schema.addFieldModal.addFieldButtonLabel": "添加字段",
+    "xpack.enterpriseSearch.schema.addFieldModal.description": "字段添加后，将无法从架构中删除。",
+    "xpack.enterpriseSearch.schema.addFieldModal.title": "添加新字段",
     "xpack.enterpriseSearch.schema.addFieldModal.fieldNameNote.correct": "字段名称只能包含小写字母、数字和下划线",
-    "xpack.enterpriseSearch.schema.addFieldModal.fieldNameNote.corrected": "该字段将被命名",
-    "xpack.enterpriseSearch.schema.addFieldModal.fieldNameNote.description": "字段添加后，将无法从架构中删除。",
-    "xpack.enterpriseSearch.schema.addFieldModal.fieldNameNote.title": "添加新字段",
+    "xpack.enterpriseSearch.schema.addFieldModal.fieldNameNote.corrected": "该字段将被命名 {correctedName}",
     "xpack.enterpriseSearch.schema.errorsCallout.buttonLabel": "查看错误",
     "xpack.enterpriseSearch.schema.errorsTable.control.review": "复查",
     "xpack.enterpriseSearch.schema.errorsTable.heading.error": "错误",
